backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asttokens==3.0.1
certifi==2026.4.22
click==8.3.1
cloudpickle==3.1.2
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.20
decorator==5.2.1
distro==1.9.0
et_xmlfile==2.0.0
executing==2.2.1
fastapi==0.135.1
fonttools==4.62.1
groq==1.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
imbalanced-learn==0.14.1
ipykernel==7.2.0
ipython==9.10.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
joblib==1.5.3
jupyter_client==8.8.0
jupyter_core==5.9.1
kiwisolver==1.5.0
llvmlite==0.46.0
matplotlib==3.10.8
matplotlib-inline==0.2.1
nest-asyncio==1.6.0
numba==0.64.0
numpy==2.4.3
openpyxl==3.1.5
packaging==26.0
pandas==3.0.1
parso==0.8.6
pillow==12.1.1
platformdirs==4.9.4
prompt_toolkit==3.0.52
psutil==7.2.2
pure_eval==0.2.3
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-multipart==0.0.22
pyzmq==27.1.0
scikit-learn==1.8.0
scipy==1.17.1
seaborn==0.13.2
shap==0.51.0
six==1.17.0
sklearn-compat==0.1.5
slicer==0.0.8
sniffio==1.3.1
stack-data==0.6.3
starlette==1.0.0
threadpoolctl==3.6.0
tornado==6.5.5
tqdm==4.67.3
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uvicorn==0.42.0
wcwidth==0.6.0
xgboost==3.2.0
passlib[bcrypt]==1.7.4
python-jose==3.3.0
python-dotenv==1.0.1
email-validator==2.3.0
bcrypt==4.0.1
pydantic[email]==2.12.5
motor==3.7.1
pymongo==4.17.0
